Update OAPMessage.py
There is a problem with the initialization of self.value as a TLVShort and self.sample_count as a TLVByte for digital channels in the Sensor class. The digital channels send 2 bytes for sample count and 1 byte for value that is why OAPMessage.TLV.parse_value(val_str) fails to parse the value tag () as it runs out of bounds.